requirements.txt
PyYAML==6.0.2
typer==0.12.3
pydantic==2.8.2
requests==2.32.3
pytest==8.3.2
pytest-cov==5.0.0
